Adds logging & implements proper INFO messages.
Now, we use the Python `logging` interface to output messages to
stdout. The formatting is still pending, as it'll eventually write
to a file, but for now is simpler for cleaner stdoutput.

INFO messages are properly transmitted on node join, so that
successor nodes can be set up. Similarly, the first part of the
`.stabilize()` sequence is working. The next step is the NOTIFY
messages to be properly sent / parsed / applied.