Fixes notify messages, improves logging, better hashing.
In this change, there are plenty of useful INFO messages being
logged now, so the default logging setting can be safely set to
INFO for general-purpose running of the scripts.

NOTIFY messages are properly parsed and communicated. With this
came proper hashing / unhashing of message data and adds a
hash parameter in the relevant messages (like JOIN).

In general, there is a TODO list now, improved error-handling,
and code cleanup.