Adds remote successor lookup support.
This adds two message types: LOOKUP and LOOKUP-RESP. These are
used to remotely look up successors for nodes. These will likely
be used in the future for proper routing of messages as well, and
will continue to use the recursive lookup approach.

This also fixes several internal bugs (for example, like in
`FingerTable.find_predecessor`), and also uses the improved packing
interface for addresses and hashes.