Fixes INFO request bug.
The specific cause of this bug was that if the predecessor wasn't
set, we fall back on US being the predecessor, which is strictly
incorrect. Our predecessor in that case is the closest node
preceding us. But, since this only occurs for the first node, we
set it to be our successor, which is guaranteed to exist.

TODO: Actually perform the `closest_preceding` lookup?