Improves partial binding and adds preliminary lookup support.
The system was stress-tested with 100 nodes and everything passes.
With 200 nodes (in a single process), the number of open sockets
gets too large, resulting in a "Too many open files" exception.

Previously, this was occurring with slightly over 50 nodes, so this
is a huge improvement. :) now for finger tables for faster
stabilization! It will result in more connections per-node, but
is a necessary evil for better performance. Currently,
stabilization takes _really_ long with increased node count.